Dashboard menu Done

+++ src/Pages/Dashboard/SingleEnrolledCourse/SingleEnrolledCourse.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import useEnrolledCourse from '../../../Hooks/useEnrolledCourse';
+import Swal from 'sweetalert2';
 
 const SingleEnrolledCourse = ({ course }) => {
     const { _id, courseName, price, instructor_name, image } = course;
@@ -8,21 +9,37 @@ const SingleEnrolledCourse = ({ course }) => {
 
 
     const handleDelete = id => {
-        fetch(`http://localhost:5000/my-course/${id}`,{
-            method: 'DELETE',
-        })
-        .then(res => res.json())
-        .then(data => {
-            refetch();
-            console.log(data);
-        })
-    }
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You want to be Delete  this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/my-course/${id}`, {
+                    method: "DELETE"
+                })
+                    .then((res) => res.json())
+                    .then((data) => {
+                        refetch();
+                        if (data.deletedCount > 0) {
+                            Swal.fire("Deleted!", "Products deleted.", "success");
+                        }
+                    });
+            }
+        });
+    };
+
+
     return (
         <div>
             <table className="table w-full">
 
                 <tbody>
-                    <tr className='bg-[#EEEEEE] rounded-md grid grid-cols-5 mb-5 justify-items-center items-center shadow-lg '>
+                    <tr className='bg-[#EEEEEE] rounded-md grid md:grid-cols-5 mb-5 justify-items-center items-center shadow-lg '>
                         <td>
                             <button onClick={() => handleDelete(_id)} className="btn btn-circle btn-error">
                                 <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M6 18L18 6M6 6l12 12" /></svg>
@@ -34,7 +51,7 @@ const SingleEnrolledCourse = ({ course }) => {
                         </td>
 
                         <td>
-                           <p className='text-xl font-serif font-semibold'>{courseName}</p>
+                            <p className='text-xl font-serif font-semibold'>{courseName}</p>
                         </td>
 
                         <td>
